+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - New #126: Add `nanostarAmount` field to `StarTransaction` type.
 - New #126: Add `TransactionPartnerAffiliateProgram` type.
 - New #126: Add `AffiliateInfo` type.
+- New #126: Add `affiliate` field to `TransactionPartnerUser` type.
 
 ## 0.4.6 November 18, 2024